Trivial fix to load_classes in correct order
The order in which the files are loaded is not deterministic. A 32bit machine
and a 64bit machine where returning files in a different order, which caused
loading of classes (with slices) to fail.

"Dir.glob/Dir[] results are determined by the underlying call to glob. In
general (under posix) the glob function sorts in ascending ASCII order
but this is not always going to be the same as alphabetical order so
any ruby code relying on this is bugged." - ara.t.howard gmail.com