[merb-core] Fixing request parameter handling for on ASCII encodings.
Merb rely on Rack for parameter encoding. ENV from Rack is in ASCII by
default. It is possible to cast the whole Rack ENV to UTF-8 but there is
problem with Merb::Parse#unescape which turns UTF-8 strings back to the
ASCII.

This patch allows to pass encoding to the Merb::Parse#unescape to force
encoding. This is only possible on Ruby 1.9. When on Ruby 1.9. and
encoding parameter is not passed unescape defaults to the
Encoding.default_internal which is nil by default.

If all attempts to get non-nil value fails no force encoding is called
and all strings are untouched.

So for uscase where you want your application to work in UTF-8
exclusively you can set Encoding.default_internal = 'UTF-8' and all your
parameters will be in UTF-8.