[merb-core] Fixing multipart parsing for identical param names
When posting multipart form with fields having the same name, fields was
only having value of the first occurence. This makes problems for
checkbox helper which add hidden before the checkbox to handle on/off
states of the checkbox.

This patch prevents multipart parser to parse files 2x which cause
problems but allows parsing non-files multiple times to fix the problem.

[#1304 state:resolved]