[merb-core, merb-slices] WIP: Moving from Extlib to ActiveSupport 3.
First try to make Merb running with ActiveSupport. Be aware that this
commit seriously break Merb. I managed to update require statements
and extract some of the Extlib classes which didn't have equivalent
in the ActiveSupport.

Interestingly enough most of the specs for the merb-core still pass
despite that there must be serious bugs and the Merb is able just
about to boot but doesn't handle single request without exception.