[merb-core] Switch specs to RSpec 2.x
This is a breaking change (at least it will be for a while) since this
initial conversion is aiming at running without errors at the rspec level.
Correctness of the specs still has to be verified.

The 1.0 specs have been deleted, mostly since maintaining two sets of specs
is beyond the current resources.

For now, just a simple "spec" task is provided.